requirements.txt
Django==1.10.6
django-cors-headers==1.3.1
djangorestframework==3.6.2
djangorestframework-jwt==1.8.0
PyJWT==1.4.2
six==1.10.0
